requirements.txt
# Django and Core Dependencies
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# Image Processing
Pillow==10.0.0

# Utilities
asgiref==3.6.0
sqlparse==0.4.4

# Markdown Processing
markdown==3.5.1
beautifulsoup4==4.12.2

